Namespace Changes:
The reason for this fork and for the commit is that I need to use both AnyEvent::ZeroMQ and ZeroMQ perl modules.  Each of these has a ZeroMQ::Raw namespace and they overwrite each other.

As a user, this totally sucks.  We should be able to use both and not have conflicts so the direction we (our team) decided to take was to fork/update this module and change the namespace from

ZeroMQ::Raw to AnyEvent::ZeroMQ::Raw

We will also fork and update the zeromq-raw project as well to make it aware of the namespace change.

In the end, both libraries are cool.  We want to use both and this is our way of solving the problem.

We will contact the author of this module and get his thoughts on the matter but until then... with tight deadlines... we will fork first and ask later :).

We have also added all the necessary files so we can actually build this module once we pull it from github.  I just copied over the files that were already in the release on cpan.